% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_sl}
\alias{classify_sl}
\title{Classify behaviour from an accelerometer axis}
\usage{
classify_sl(
  data,
  axis,
  threshold,
  name = "Behaviour",
  below = "Standing",
  above = "Lying",
  min_run = 0,
  comparison = "<"
)
}
\arguments{
\item{data}{A data frame containing accelerometer data.}

\item{axis}{The name of a column in \code{data} to be used for this classification.}

\item{threshold}{The accelerometer threshold to use for classification.}

\item{name}{(optional) The name for the new column with the classified behaviour. Defaults to "Behaviour".}

\item{below}{(optional) The behaviour to assign when below the threshold. Defaults to "Standing".}

\item{above}{(optional) The behaviour to assign when above the threshold. Defaults to "Lying".}

\item{min_run}{(optional) The shortest consecutive sequence to keep. Defaults to 0 (no smoothing).}

\item{comparison}{(optional) One of \code{"<"} or \code{"<="}, determines whether values on the threshold are
assigned to \code{above} or \code{below}. Defaults to "<".}
}
\value{
a copy of \code{data} with a new column \code{data[[name]]} containing the accelerometer-derived behaviours.
}
\description{
Use this function to convert a single accelerometer axis into a behavioural classification,
usually Standing, Lying.
This classifier uses a simple decision rule: it chooses one of two behaviours based on whether
the acceleration measured by the axis is below or above a user-supplied threshold.
}
\examples{
goats |> dplyr::group_by(Animal) |> classify_sl("X", -0.60)

}
