% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_slr}
\alias{classify_slr}
\title{Classify behaviour from two accelerometer axes}
\usage{
classify_slr(
  data,
  axis1,
  threshold1,
  axis2,
  threshold2,
  name = "Behaviour",
  below1 = "Standing",
  below2 = "Lying Left",
  above2 = "Lying Right",
  min_run = 0,
  comparison1 = "<",
  comparison2 = "<"
)
}
\arguments{
\item{data}{A data frame containing accelerometer data.}

\item{axis1}{The name of a column in \code{data} to be used for the SL classification.}

\item{threshold1}{The accelerometer threshold to use for the SL classification.}

\item{axis2}{The name of a column in \code{data} to be used for the LR classification.}

\item{threshold2}{The accelerometer threshold to use for the LR classification.}

\item{name}{(optional) The name for the new column with the classified behaviour. Defaults to "Behaviour".}

\item{below1}{(optional) The name for the first behaviour. Defaults to "Standing".}

\item{below2}{(optional) The name for the second behaviour. Defaults to "Lying Left"}

\item{above2}{(optional) The name for the third behaviour. Defaults to "Lying Right"}

\item{min_run}{(optional) The shortest consecutive sequence to keep. Defaults to 0 (no smoothing).}

\item{comparison1}{(optional) one of \code{"<"}, \code{"<="}, \code{">"}, or \code{">="},
determines the comparison between the first axis and first threshold. Defaults to "<".}

\item{comparison2}{(optional) one of \code{"<"} or \code{"<="},
determines whether values on the threshold are assigned to \code{below2} or \code{above2}. Defaults to "<".}
}
\value{
a copy of \code{data} with a new column \code{data[[name]]} containing the accelerometer-derived behaviours.
}
\description{
Use this function to convert two accelerometer axes into a behavioural classification,
usually Standing, Lying Left, Lying Right.
This classifier uses a two-part decision rule: it chooses the first behaviour (S vs L or R) based
on the first axis and threshold, or it uses the second axis and threshold
to choose between the second and third behaviour (L vs R).
}
\examples{
classify_slr(goats, "X", -0.6, "Z", -0.65)

}
