% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robo.R
\name{datasets}
\alias{datasets}
\alias{cows}
\alias{goats}
\title{Example datasets}
\format{
\describe{
\item{\code{Animal}}{Animal ID.}
\item{\code{Time}}{Time and date.}
\item{\code{X}}{x-axis acceleration (g) (only in \code{goats}).}
\item{\code{Y}}{y-axis acceleration (g).}
\item{\code{Z}}{z-axis acceleration (g).}
}
}
\description{
Examples of (cleaned) HOBO accelerometer data. One data set, \code{cows}, is from dairy
cows, with horizontally mounted devices, and with y-axis and z-axis recorded.
The second, \code{goats}, is from dairy goats, with vertically mounted devices, and
all three axes recorded.

Both data sets have been de-identified. They are suitable for learning to use
the robo package, but should be used with caution since the de-identification
may have made some aspects of the data unrealistic.
}
