% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{group_daily}
\alias{group_daily}
\title{Group by day}
\usage{
group_daily(data, time = "Time", name = "Day")
}
\arguments{
\item{data}{A data frame to be grouped by day.}

\item{time}{(optional) The name of the time column. Defaults to \code{"Time"}.}

\item{name}{(optional) The name to give to the day column. Defaults to \code{"Day"}.}
}
\value{
A grouped data frame, with a new column for day, and grouped by day.
}
\description{
Group a data frame by day, after adding a column for the day.
If the data frame was already grouped, those previous groups are preserved.
}
\examples{
cows |> dplyr::group_by(Animal) |> group_daily()

}
