% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{group_hourly}
\alias{group_hourly}
\title{Group by hour}
\usage{
group_hourly(data, time = "Time", name_day = "Day", name_hour = "Hour")
}
\arguments{
\item{data}{A data frame to be grouped by hour.}

\item{time}{(optional) The name of the time column. Defaults to \code{"Time"}.}

\item{name_day}{(optional) The name to give to the day column. Defaults to \code{"Day"}.}

\item{name_hour}{(optional) The name to give to the hour column. Defaults to \code{"Hour"}.}
}
\value{
A grouped data frame, with a columns for day and hour, and grouped by day and hour.
}
\description{
Group a data frame by day and hour, after adding columns for day and hour.
If the data frame was already grouped, those previous groups are preserved.
}
\examples{
cows |> dplyr::group_by(Animal) |> group_hourly()

}
