% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hobo_example}
\alias{hobo_example}
\title{Get path to example files}
\usage{
hobo_example(animal = c("cows", "goats"))
}
\arguments{
\item{animal}{Which set of example files to get, either "cows" or "goats".}
}
\value{
A path to the example files as a character string.
}
\description{
Get the local path to the example files bundled with robo.
}
\examples{
hobo_example("goats")

}
