% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_hobo}
\alias{read_hobo}
\title{Read a single HOBO CSV}
\usage{
read_hobo(file, tz)
}
\arguments{
\item{file}{Path to the HOBO CSV file.}

\item{tz}{(optional) Time zone to use for the time column.
If this argument is missing, it will be inferred from the HOBO column heading.}
}
\value{
A data frame:

\describe{
\item{\code{Animal}}{Animal ID.}
\item{\code{Time}}{Time and date.}
\item{\code{X}, \code{Y}, \code{Z}}{Acceleration (g) for axes present in the data.}
}
}
\description{
Read an exported HOBO CSV into R as a data frame.
See \code{vignette("tutorial")} for the expected export format.
Animal ID is inferred from the "Plot Title" header.
}
\examples{
file_cow01 <- file.path(hobo_example("cows"), "cow01.csv")
cow01 <- read_hobo(file_cow01)

}
