% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_hobo_dir}
\alias{read_hobo_dir}
\title{Read HOBO CSVs from a directory}
\usage{
read_hobo_dir(dir, tz)
}
\arguments{
\item{dir}{Path to the directory.}

\item{tz}{(optional) Time zone to use for the time column.
If this argument is missing, it will be inferred from the HOBO column heading.}
}
\value{
A single data frame containing data from the whole directory.
}
\description{
Read every CSV in a directory using \code{\link[=read_hobo]{read_hobo()}} and join them into a single data frame.
}
\examples{
cows_path <- hobo_example("cows")
cows_raw <- read_hobo_dir(cows_path, tz="America/New_York")

}
\seealso{
read_hobo
}
