% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_behaviour}
\alias{smooth_behaviour}
\alias{smooth_behavior}
\title{Apply smoothing to behaviour data}
\usage{
smooth_behaviour(data, name_in = "Behaviour", name_out = name_in, min_run = 2)

smooth_behavior(data, name_in = "Behaviour", name_out = name_in, min_run = 2)
}
\arguments{
\item{data}{A data frame containing classified behaviour data.}

\item{name_in}{(optional) The column name containing classified behaviours. Defaults to "Behaviour".}

\item{name_out}{(optional) The column name for smoothed behaviour. If this isn't supplied, the function
will overwrite the original column.}

\item{min_run}{The minimum number of consecutive observations to trust.}
}
\value{
The input data frame, with a column of smoothed behaviour data.
}
\description{
Short bouts are often noise in the accelerometer data rather than genuine behaviour changes.
Smoothing out short runs, especially singletons, can improve the overall accuracy of behaviour classifications.
This function uses run-length encoding to replace short bouts with the previously classified behaviour.
This function acts on data frames --- to apply the same smoother to a vector use \code{\link[=smooth_points]{smooth_points()}}.
}
\examples{
cows |> dplyr::group_by(Animal) |> classify_sl("Y", -0.65) |> smooth_behaviour(name_out="Smoothed") 

}
