% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_bouts}
\alias{smooth_bouts}
\title{Apply smoothing to bouts data}
\usage{
smooth_bouts(data, min_length = 2, behaviour = attr(data, "behaviour"))
}
\arguments{
\item{data}{A data frame where each row is a bout.
The data frame should have \code{Start}, \code{End} and \code{Length} columns.}

\item{min_length}{Minimum bout length in seconds.}

\item{behaviour}{(optional) The name of the behaviour column. Defaults to \code{"Behaviour"}.}
}
\value{
A data frame with the same columns as \code{data} but with all bouts having
duration at least \code{min_length} seconds.
}
\description{
Take a data frame where rows represent bouts, and apply a smoothing algorithm
to ensure all bouts have a minimum length.
\itemize{
\item Long bouts are kept.
\item Long series of short bouts are merged, with the behaviour decided by voting.
\item Short series of short bouts, or isolated short bouts, are merged into their neighbours.
}

Any consecutive bouts of the same behaviour are merged. If the total duration within a group is
less than the minimum bout length, no bouts are returned for that group.
}
\examples{
\donttest{
bouts <- goats |> dplyr::group_by(Animal) |> classify_sl("X", -0.60) |> group_daily() |> bouts()
smooth_bouts(bouts, 300)
}

}
