% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_points}
\alias{smooth_points}
\title{Apply smoothing to a vector}
\usage{
smooth_points(x, min_run = 2)
}
\arguments{
\item{x}{A vector of behaviours.}

\item{min_run}{The minimum number of consecutive observations to trust.}
}
\value{
A vector the same length as \code{x}, with smoothing applied.
}
\description{
Apply the smoother from \code{\link[=smooth_behaviour]{smooth_behaviour()}} to a vector of behaviours.
}
\examples{
bvr <- c("a", "a", "a", "b", "a", "a", "a", "b", "b", "a")
smooth_points(bvr, min_run=2)
smooth_points(bvr, min_run=3)

}
\seealso{
smooth_behaviour
}
