% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise_bouts}
\alias{summarise_bouts}
\alias{summarize_bouts}
\title{Summarise grouped bout data}
\usage{
summarise_bouts(data, behaviour = "Behaviour")

summarize_bouts(data, behaviour = "Behaviour")
}
\arguments{
\item{data}{A grouped data frame where each row is a bout.
The data frame should have a \code{Length} column.}

\item{behaviour}{(optional) The name of the behaviour column. Defaults to `"Behaviour"``.}
}
\value{
An ungrouped data frame in "long" form, with a row for each combination
of group and behaviour, with the number and duration of that kind of bout for that group.
}
\description{
Take a grouped data frame produced by \code{\link[=bouts]{bouts()}}, and summarise the number and duration of bouts in each group.
}
\examples{
bouts <- goats |> dplyr::group_by(Animal) |> classify_sl("X", -0.60) |> group_daily() |> bouts()
summarise_bouts(bouts)

}
\seealso{
bouts
}
