% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise_time}
\alias{summarise_time}
\alias{summarize_time}
\title{Summarise data grouped by time}
\usage{
summarise_time(
  data,
  behaviour = "Behaviour",
  time = "Time",
  units = c("hours", "mins", "secs"),
  max_step = 5
)

summarize_time(
  data,
  behaviour = "Behaviour",
  time = "Time",
  units = c("hours", "mins", "secs"),
  max_step = 5
)
}
\arguments{
\item{data}{A data frame grouped by time period.}

\item{behaviour}{(optional) The name of the behaviour column. Defaults to \code{"Behaviour"}.}

\item{time}{(optional) The name of the time column. Defaults to \code{"Time"}.}

\item{units}{(optional) Units for bout durations, one of \code{"hours"}, \code{"mins"}, \code{"secs"}.
The default is \code{"hours"}.}

\item{max_step}{(optional) The maximum duration, in seconds, to attribute to an observation.
Defaults to 5.}
}
\value{
An ungrouped data frame with any grouping columns,
a column for each behaviour duration, and a column for total duration.
}
\description{
Take a grouped data frame and summarise the duration of behaviours in each group.

Durations for a given instantaneous observation are assumed to span the time from the observation
itself to the time of the next observation, unless that duration exceeds \code{max_step} seconds.
The duration of the final observation in a group is assumed to be the modal time step within the group.
}
\examples{
goats |> dplyr::group_by(Animal) |> classify_sl("X", -0.60) |> group_daily() |> summarise_time()

}
